Add scroll animation to projects section.
